[Embedding] Split GPUHashTable into standalone file and refactor build dependency.
