+++ tensorflow/core/BUILD
@@ -3005,8 +3005,8 @@ tf_cuda_library(
             "util/env_var.cc",
             "framework/embedding/batch.cu.cc",
             "framework/embedding/batch.h",
-            "framework/embedding/embedding_var.cu.cc",
-            "framework/embedding/embedding_var.h"
+            "framework/embedding/gpu_hash_table.cu.cc",
+            "framework/embedding/gpu_hash_table.h"
         ],
     ) + select({
         "//tensorflow:windows": [],
@@ -6037,11 +6037,11 @@ cuda_library(
     name = "embedding_gpu",
     hdrs = [
         "framework/embedding/batch.h",
-        "framework/embedding/embedding_var.h"
+        "framework/embedding/gpu_hash_table.h"
     ],
     srcs = FRAMEWORK_INTERNAL_PRIVATE_HEADERS + [
         "framework/embedding/batch.cu.cc",
-        "framework/embedding/embedding_var.cu.cc"
+        "framework/embedding/gpu_hash_table.cu.cc"
     ],
     copts = ["-g"] + if_cuda(["-DNV_CUDNN_DISABLE_EXCEPTION"])
         + if_cuda(["-DGOOGLE_CUDA=1"]) +
@@ -6060,10 +6060,8 @@ cuda_library(
     alwayslink = 1,
     deps = ["//tensorflow/stream_executor/cuda:cudart_stub",
 	    ":stream_executor",
-        "@sparsehash_c11//:dense_hash_map",
-        "@libcuckoo//:libcuckoo",
-        "@com_github_google_leveldb//:leveldb",
-        ] + select({"//tensorflow:with_gpu_ev_support":
-                 ["@cuCollections//:cuco_hash_table"],
-                 "//conditions:default": []})
+            "@sparsehash_c11//:dense_hash_map",
+            "@libcuckoo//:libcuckoo",
+            "@com_github_google_leveldb//:leveldb",
+            "@cuCollections//:cuco_hash_table"],
 )
