+++ tensorflow/core/framework/embedding/batch.cu.cc
@@ -1,5 +1,19 @@
+/* Copyright 2022 The DeepRec Authors. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+======================================================================*/
+
 #if GOOGLE_CUDA
-#if !TENSORFLOW_USE_GPU_EV
 
 #include "tensorflow/core/framework/embedding/batch.h"
 #include "tensorflow/core/framework/register_types.h"
@@ -80,5 +94,4 @@ __global__ void CopyEmbedding(V** batch, V** batch_data_space,
 TF_CALL_REAL_NUMBER_TYPES(REGISTER_KERNELS_ALL_INDEX)
 #undef REGISTER_KERNELS_ALL_INDEX
 }  // namespace tensorflow
-#endif  // TENSORFLOW_USE_GPU_EV
 #endif  // GOOGLE_CUDA
