+++ tensorflow/core/framework/embedding/batch.h
@@ -1,9 +1,22 @@
+/* Copyright 2022 The DeepRec Authors. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+======================================================================*/
+
 #ifndef TENSORFLOW_CORE_FRAMEWORK_EMBEDDING_BATCH_
 #define TENSORFLOW_CORE_FRAMEWORK_EMBEDDING_BATCH_
 
 #if GOOGLE_CUDA
-#if !TENSORFLOW_USE_GPU_EV
-
 namespace tensorflow {
 
 template<class V>
@@ -25,7 +38,6 @@ __global__ void CopyEmbedding(V** batch, V** batch_data_space,
 
 }  // namespace tensorflow
 
-#endif  // TENSORFLOW_USE_GPU_EV
 #endif  // GOOGLE_CUDA
 
 #endif  // TENSORFLOW_CORE_FRAMEWORK_EMBEDDING_BATCH_
